Proxy improvements
- Removed abstract Proxy class
- Renamed Http -> HttpProxy and NetworkInterface ->
NetworkInterfaceProxy
- Removed ProxyInterface::getId() method
- Fixed namespace bug in NetworkInterfaceProxy
- Changed accepted instance of proxy class from Proxy to ProxyInterface
- Added examples of using proxies in README.md
