--- src/Erpk/Harvester/Client/Proxy/Proxy.php
@@ -1,10 +1,0 @@
-<?php
-namespace Erpk\Harvester\Client\Proxy;
-
-abstract class Proxy
-{
-    public function getId()
-    {
-        return $this->id;
-    }
-}
